update chombo entry
