+++ src/pages/league/[id]/index.tsx
@@ -23,7 +23,7 @@ import RankedEventDetails, {
 } from '../../../components/display/RankedEventDetails';
 import MatchEntryDialog, {
   RankedMatch,
-} from '../../../components/display/MatchEntryDialog';
+} from '../../../components/matchEntry/MatchEntryDialog';
 import Leaderboard from '../../../components/display/Leaderboad';
 import { PersonalStats } from '../../../components/display/PersonalStats';
 import ScoreHistory from '../../../components/display/ScoreHistory';
